requirements.txt
duckdb
click>=8
networkx>=3
jinja2>=3
python-dotenv>=1
pyyaml>=6
werkzeug>=3